Moved importBoard to file. Added missing root to mkdir and listFile in io.native
